+++ assets/js/blocks/products/all-products/block.js
@@ -4,14 +4,14 @@
 import { Component } from 'react';
 import PropTypes from 'prop-types';
 import ProductListContainer from '@woocommerce/base-components/product-list/container';
-import { InnerBlockParentNameProvider } from '@woocommerce/base-context/inner-block-parent-name-context';
+import { InnerBlockConfigurationProvider } from '@woocommerce/base-context/inner-block-configuration-context';
 import { ProductLayoutContextProvider } from '@woocommerce/base-context/product-layout-context';
 
 const layoutContextConfig = {
 	layoutStyleClassPrefix: 'wc-block-grid',
 };
 
-const parentBlockConfig = { blockName: 'woocommerce/all-products' };
+const parentBlockConfig = { parentName: 'woocommerce/all-products' };
 
 /**
  * The All Products Block. @todo
@@ -33,14 +33,14 @@ class Block extends Component {
 		 * wc-block-{$this->block_name},
 		 */
 		return (
-			<InnerBlockParentNameProvider value={ parentBlockConfig }>
+			<InnerBlockConfigurationProvider value={ parentBlockConfig }>
 				<ProductLayoutContextProvider value={ layoutContextConfig }>
 					<ProductListContainer
 						attributes={ attributes }
 						urlParameterSuffix={ urlParameterSuffix }
 					/>
 				</ProductLayoutContextProvider>
-			</InnerBlockParentNameProvider>
+			</InnerBlockConfigurationProvider>
 		);
 	}
 }
